Implement gouraud shading and phong shading

+++ culling_clipping.cpp
@@ -3,15 +3,15 @@
 #include <cmath>
 using namespace std;
 
-double backfaceCulling(vector3f v1, vector3f v2, vector3f v3, vector3f lightdir)
+float backfaceCulling(vector3f v1, vector3f v2, vector3f v3, vector3f lightdir)
 {
     vector3f va = v2 - v1;
     vector3f vb = v3 - v1;
     vector3f vc = cross(va, vb);
-    double mod = (double)(vc.x*vc.x + vc.y*vc.y + vc.z*vc.z);
-    vector3f vn(((double)vc.x)/sqrt(mod), ((double)vc.y)/sqrt(mod), ((double)vc.z)/sqrt(mod));
+    float mod = (float)(vc.x*vc.x + vc.y*vc.y + vc.z*vc.z);
+    vector3f vn(((float)vc.x)/sqrt(mod), ((float)vc.y)/sqrt(mod), ((float)vc.z)/sqrt(mod));
 
-    double intensity = dot(vn, lightdir);
+    float intensity = dot(vn, lightdir);
 //    cout << "in function: " << vn.x << " " << vn.y << " " << vn.z << endl;
     return intensity;
 }
