+++ rasterizer.h
@@ -9,6 +9,6 @@
 void bresenham(vector2d v1, vector2d v2, TGAImage &image, TGAColor color);
 
 void scanLine(vector2d v1, vector2d v2, vector2d v3, TGAImage &image, TGAColor color);
-void baryCentric(vector3d v1, vector3d v2, vector3d v3, TGAImage &image, TGAColor color, double intensity, int* zbuffer, model m, int cnt);
+void baryCentric(vector3d v1, vector3d v2, vector3d v3, TGAImage &image, TGAColor color, float intensity, int* zbuffer, model m, int cnt);
 
 #endif // RASTERIZER_H_INCLUDED
